requirements.txt
affine==2.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.54.1
GDAL==3.10.0
GeoAlchemy2==0.15.2
geopandas==1.0.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
rasterio==1.4.1
shapely==2.0.6
six==1.16.0
SQLAlchemy==2.0.36
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
zipp==3.20.2
